Add files via upload
The Titanic dataset is downloaded from Kaggle site and the Random Forest classifier, Logistic regression, and Decision Tree classifiers are built in Python using Sklearn library and are visualized using seaborn, matplotlib. 
This Analysis is used to predict the survival rate of passengers by using certain attributes. 2 fold validation is used in order to create validation, training sets and can be used for tuning hyper-parameters in order to avoid overfitting.

The various models are evaluated using Confusion matrix, Decile analysis (cumulative lift across the depth of classification). Misclassification ratios, precision, and recall are used to understand the error rate.

ROC Curve and AUC curve are created and visualized to determine the parameters such as max_depth, max_leaf_size, number of trees in Random Forest.